valgrind: Mark excess code space as unaddressable
When PCRE compiles a pattern, the required space is first estimated and this
amount of memory is then allocated. For some patterns, the actually required
code space is less than that, which means that the end of the code space is
unused. This patch marks that region as unaddressable, so that valgrind will
signal an error when it is accessed.

git-svn-id: svn://vcs.exim.org/pcre/code/trunk@1152 2f5784b3-3f2a-0410-8824-cb99058d5e15